requirements.txt
discord.py>=2.3.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
alembic>=1.12.0